in do_browser(), make old_selected a size_t too, to match selected
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@3315 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/browser.c
@@ -114,7 +114,7 @@ char *do_browser(char *path, DIR *dir)
 		/* If we clicked in the edit window, we probably clicked
 		 * on a file. */
 		if (wenclose(edit, mevent.y, mevent.x)) {
-		    int old_selected = selected;
+		    size_t old_selected = selected;
 
 		    /* Subtract out the size of topwin. */
 		    mevent.y -= 2 - no_more_space();
